refactor: make download detail an inspector
