+++ Harbor/Views/DownloadDetailView.swift
@@ -4,12 +4,8 @@ struct DownloadDetailView: View {
     let center: DownloadCenter
 
     var body: some View {
-        Group {
-            if let item = center.selectedDownload {
-                DownloadInspectorContent(item: item, center: center)
-            } else {
-                EmptyDownloadDetailView(addDownload: center.presentAddSheet)
-            }
+        if let item = center.selectedDownload {
+            DownloadInspectorContent(item: item, center: center)
         }
     }
 }
@@ -99,24 +95,6 @@ private struct DownloadInspectorContent: View {
 
     private func copySourceURL() {
         center.copySourceURL(id: item.id)
-    }
-}
-
-private struct EmptyDownloadDetailView: View {
-    let addDownload: () -> Void
-
-    var body: some View {
-        ContentUnavailableView {
-            Label("Select a Download", systemImage: "sidebar.right")
-        } description: {
-            Text("Choose any row to inspect progress, speed, file location, and recovery actions.")
-        } actions: {
-            Button(action: addDownload) {
-                Label("Add Download", systemImage: "plus")
-            }
-            .buttonStyle(LiquidPillButtonStyle(prominent: true))
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
     }
 }
 
@@ -1007,44 +985,6 @@ private struct DownloadDetailSection<Content: View>: View {
             content()
         }
         .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-private struct LiquidPillButtonStyle: ButtonStyle {
-    let prominent: Bool
-
-    func makeBody(configuration: Configuration) -> some View {
-        let label = configuration.label
-            .font(.callout.weight(.medium))
-            .lineLimit(1)
-            .padding(.horizontal, 14)
-            .padding(.vertical, 7)
-            .frame(minHeight: 32)
-            .foregroundStyle(prominent ? Color.white : Color.secondary)
-            .opacity(configuration.isPressed ? 0.78 : 1)
-            .scaleEffect(configuration.isPressed ? 0.98 : 1)
-
-        if #available(macOS 26, *) {
-            label
-                .glassEffect(
-                    prominent ? .regular.tint(.accentColor).interactive() : .regular.interactive(),
-                    in: .rect(cornerRadius: 16)
-                )
-                .contentShape(.rect(cornerRadius: 16))
-        } else {
-            label
-                .background(
-                    prominent ? Color.accentColor.opacity(0.16) : Color.secondary.opacity(0.08),
-                    in: RoundedRectangle(cornerRadius: 16, style: .continuous)
-                )
-                .overlay {
-                    RoundedRectangle(cornerRadius: 16, style: .continuous)
-                        .strokeBorder(
-                            prominent ? Color.accentColor.opacity(0.24) : Color.secondary.opacity(0.16)
-                        )
-                }
-                .contentShape(.rect(cornerRadius: 16))
-        }
     }
 }
